Hide the internal API more, and move the tests
